Give the pacer's burst session a lifetime, and an object to live in
The pacer held nine fields whose only shared property was that clear() had to
remember all of them - and forget the learned overhead beside them. Move the
six session-scoped ones into a CaptureAvailablePacingSession that starts on
the burst's first capture and is dropped when the queue drains, so clear() is
one line and cannot reset the wrong thing. The session owns the admitted-work
FIFO and the backlog clock derived from it, rather than exposing both for the
pacer to mutate: admit() reads its delay and work back off the decision it
queues, so the clock cannot advance by anything that decision was not built
on. sessionDraftOverheadMs is gone as redundant - it was always equal to the
overhead already snapshotted into the pacing prediction.

Two observability fixes. The burst-session id was a per-pacer ordinal that
restarted at 0 whenever the draft pipeline was recreated, so two camera
sessions merged into one whenever neither had drained; it is now the session's
creation uptime, which a later session can never repeat. And the exporter's
"totalShotCount" reported the per-session count, not a total - it is
sessionShotCount now, and a number rather than "#"-prefixed text.

Drop mandatoryReserveUpperBoundMs, which reached the metrics store, the
database, and the exporter to classify one log line's severity. Its removal
takes DraftSequenceEstimate with it: the point sum was all that was left, so
estimateDraftSequenceMs returns it directly. The pacing log no longer
escalates to PLog.e when the RESERVED tail outgrows the budget.

+++ DraftSequenceExecutionProfiler.kt
@@ -292,7 +292,7 @@ private class MetricsRecorder(
         captureMetrics.draftSequenceMetrics = draftSequenceMetrics
     }
 
-    fun onDraftStart(gatingPacingDecision: CaptureAvailablePacingDecision?, pacerSessionId: Int) {
+    fun onDraftStart(gatingPacingDecision: CaptureAvailablePacingDecision?, pacerSessionId: Long) {
         synchronized(draftSequenceMetrics) {
             draftSequenceMetrics.draftStartUptimeMs = SystemClock.uptimeMillis()
             draftSequenceMetrics.pacerSessionId = pacerSessionId
